# docs/swagger.yaml
basePath: /api
definitions:
...
  api.Success:
    properties:
      code:
        description: 状态码
        example: 200
        type: integer
      extime:
        description: 耗时(毫秒)
        example: 70
        type: integer
      payload:
        description: 内容
    required:
    - code
    - extime
    type: object
...
info:
  contact:
    url: https://www.jkstack.com
  description: jkagent management server.
  title: agent-server
  version: TODO
paths:
...
  /file/{id}/upload:
    post:
      operationId: /api/file/upload
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      - description: 保存路径
        in: formData
        name: dir
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: md5校验码
        in: formData
        name: md5
        type: string
      - default: 644
        description: 文件权限（8进制）
        in: formData
        name: mod
        type: integer
      - description: 文件所属用户
        in: formData
        name: own_user
        type: string
      - description: 文件所属分组
        in: formData
        name: own_group
        type: string
      - default: 60
        description: 超时时间
        in: formData
        name: timeout
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
      summary: 上传文件
      tags:
      - file
  /foo/{id}:
    get:
      operationId: /api/foo
      parameters:
      - description: 节点ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Success'
      summary: 调用example类型的agent
      tags:
      - foo
...
swagger: "2.0"